+++ tests/test_format.py
@@ -46,3 +46,21 @@ def test_unknown_format(capsys):
         run('-t', 'xxx')
     out, err = capsys.readouterr()
     assert err.startswith('output format must be one of: ')
+
+
+def test_missing_source(stdin, capsys):
+    with pytest.raises(SystemExit):
+        with stdin('source,target\n,bar\n'):
+            run('convert', '-t', 'csv')
+    out, err = capsys.readouterr()
+    assert out == 'source, target, annotation\n'
+    assert err == 'missing source in line 2\n'
+
+
+def test_missing_target(stdin, capsys):
+    with pytest.raises(SystemExit):
+        with stdin("foo,\n"):
+            run('convert', '-H', '-t', 'csv')
+    out, err = capsys.readouterr()
+    assert out == ''
+    assert err == 'missing target in line 1\n'