+++ wdmapper/format/csv.py
@@ -17,50 +17,61 @@ name = 'csv'
 extension = '.csv'
 """extension of CSV files."""
 
-
-def reader(stream, header=True):
-    """Read mappings in CSV format."""
-
-    if (header):
-        header_row = None
-    else:
-        header_row = CSV_FIELDS
-
-    stream = iter(stream.readline, '')  # support streaming input
-
-    if sys.version_info[0] > 2:
-        csv_reader = csv.reader(stream, skipinitialspace=True)
-    else:
-        csv_reader = unicode_csv_reader(stream, skipinitialspace=True)
-    for row in csv_reader:
-        if header_row:
-            # TODO: simplify this
-            mapping = []
-            mapping = {k: v for k, v in zip(header_row, row) if v != ''}
-            try:
-                link = Link(mapping['source'])
-                if 'target' in mapping:
-                    link.target = mapping['target']
-                if 'annotation' in mapping:
-                    link.annotation = mapping['annotation']
-                yield link
-            except ValueError:
-                pass
-        else:
-            header_row = row
+PY3 = sys.version_info[0] == 3
 
 
-def unicode_csv_reader(unicode_csv_data, **kwargs):
+def _unicode_csv_reader(unicode_csv_data, **kwargs):
     # csv.py doesn't do Unicode; encode temporarily as UTF-8:
+    def utf_8_encoder(unicode_csv_data):
+        for line in unicode_csv_data:
+            yield line.encode('utf-8')
     csv_reader = csv.reader(utf_8_encoder(unicode_csv_data), **kwargs)
     for row in csv_reader:
         # decode UTF-8 back to Unicode, cell by cell:
         yield [unicode(cell, 'utf-8') for cell in row]
 
 
-def utf_8_encoder(unicode_csv_data):
-    for line in unicode_csv_data:
-        yield line.encode('utf-8')
+class Reader:
+    """Read Links in CSV format."""
+
+    def __init__(self, stream, header=True):
+        self.stream = stream
+        self.header = header
+
+    def links(self):
+        stream = iter(self.stream.readline, '')  # support streaming input
+
+        if PY3:
+            csv_reader = csv.reader(stream, skipinitialspace=True)
+        else:
+            csv_reader = _unicode_csv_reader(stream, skipinitialspace=True)
+
+        header_row = None if self.header else CSV_FIELDS
+        line = 0
+
+        for row in csv_reader:
+            line = line + 1
+            if header_row:
+                # TODO: simplify this
+                mapping = []
+                mapping = {k: v for k, v in zip(header_row, row) if v != ''}
+                try:
+                    try:
+                        link = Link(source=mapping['source'])
+                    except (ValueError, KeyError):
+                        raise WdmapperError('missing source in line %d' % line)
+                    if 'target' not in mapping:
+                        raise WdmapperError('missing target in line %d' % line)
+                    link.target = mapping['target']  # also normalizes
+                    if link.target == '':
+                        raise WdmapperError('missing target in line %d' % line)
+                    if 'annotation' in mapping:
+                        link.annotation = mapping['annotation']
+                    yield link
+                except ValueError:
+                    pass
+            else:
+                header_row = row
 
 
 class Writer: